pyproject.toml
[tool.poetry]
name = "notiondipity_backend"
version = "0.6.0"
description = ""
authors = ["Milos Svana <milos.svana@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
requests = "^2.28.2"
numpy = "^1.24.2"
gunicorn = { extras = ["gevent"], version = "^21.2.0" }
pyjwt = "^2.7.0"
pycryptodome = "^3.18.0"
tiktoken = "^0.5.0"
psycopg = { extras = ["pool"], version = "^3.1.10" }
gevent = "^23.9.1"
quart = "^0.19.2"
quart-cors = "^0.7.0"
hypercorn = "^0.14.4"
aiohttp = "^3.8.5"
mixpanel = "^4.10.0"
pgvector = "^0.2.3"
openai = "^1.10.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.4"
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
